feat(execute/table): expose static table package and table diff functions
This exposes the static table package to more easily produce test
tables. This is exposed publicly so other projects, such as influxdb,
can use it to test the tables they produce.

It adds a new function to the static table to extend the table by either
adding or overwriting columns. This way it is easier to create a base
table, such as one with common group keys, and then override it for each
table in a table iterator.

This also adds a diff iterator function that will accept a table
iterator, sort it, then pass it to `cmp.Diff` as a way of encapsulating
the diff workflow.
