+++ internal/execute/table/iterator.go
@@ -1,14 +1,5 @@
 package table
 
-import "github.com/influxdata/flux"
+import "github.com/influxdata/flux/execute/table"
 
-type Iterator []flux.Table
-
-func (t Iterator) Do(f func(flux.Table) error) error {
-	for _, tbl := range t {
-		if err := f(tbl); err != nil {
-			return err
-		}
-	}
-	return nil
-}
+type Iterator = table.Iterator
