+++ internal/execute/table/sort.go
@@ -2,7 +2,7 @@ package table
 
 import (
 	"github.com/influxdata/flux"
-	"github.com/influxdata/flux/execute"
+	"github.com/influxdata/flux/execute/table"
 )
 
 // Sort will read a TableIterator and produce another TableIterator
@@ -12,21 +12,5 @@ import (
 // all of the tables are read to avoid any deadlocks. Be careful
 // using this method in performance sensitive areas.
 func Sort(tables flux.TableIterator) (flux.TableIterator, error) {
-	groups := execute.NewGroupLookup()
-	if err := tables.Do(func(table flux.Table) error {
-		buffered, err := execute.CopyTable(table)
-		if err != nil {
-			return err
-		}
-		groups.Set(buffered.Key(), buffered)
-		return nil
-	}); err != nil {
-		return nil, err
-	}
-
-	var buffered []flux.Table
-	groups.Range(func(_ flux.GroupKey, value interface{}) {
-		buffered = append(buffered, value.(flux.Table))
-	})
-	return Iterator(buffered), nil
+	return table.Sort(tables)
 }
